MAINT: Rename _pls to pls to match conventions
